+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/gwo_panel.js
@@ -8,12 +8,42 @@ if (!gwaioWarInfoPanelLoaded) {
       var game = model.game();
 
       if (!game.isTutorial()) {
-        var loadoutId = game.inventory().cards()[0].id;
-        model.gwaioLoadout = ko.observable("");
-        requireGW(["cards/" + loadoutId], function (card) {
-          model.gwaioLoadout(loc(card.summarize()));
-        });
+        // War Information
+        var originSystem = game
+          .galaxy()
+          .stars()
+          [game.galaxy().origin()].system();
 
+        model.gwaioSettings = originSystem.gwaio;
+
+        if (model.gwaioSettings) {
+          model.gwaioDifficulty = loc(model.gwaioSettings.difficulty);
+          model.gwaioSize = loc(model.gwaioSettings.galaxySize);
+
+          if (model.gwaioSettings.ai) {
+            model.gwaioAI = model.gwaioSettings.ai;
+          } else model.gwaioAI = "Titans";
+
+          model.gwaioOptions = ko.observableArray([]);
+
+          if (model.gwaioSettings.factionScaling) {
+            model.gwaioOptions.push(loc("!LOC:Faction scaling"));
+          }
+          if (model.gwaioSettings.systemScaling) {
+            model.gwaioOptions.push(loc("!LOC:System scaling"));
+          }
+          if (model.gwaioSettings.easierStart) {
+            model.gwaioOptions.push(loc("!LOC:Easier start"));
+          }
+          if (model.gwaioSettings.tougherCommanders) {
+            model.gwaioOptions.push(loc("!LOC:Tougher commanders"));
+          }
+          if (game.hardcore()) {
+            model.gwaioOptions.push(loc("!LOC:Hardcore mode"));
+          }
+        }
+
+        // Player Information
         var factions = [
           "Legonis Machina",
           "Foundation",
@@ -24,62 +54,42 @@ if (!gwaioWarInfoPanelLoaded) {
         var factionIndex = game.inventory().getTag("global", "playerFaction");
         model.gwaioFactionName = factions[factionIndex];
 
-        var stats = game.stats();
-        model.gwaioStats = ko.observable(
-          stats.turns() + "/" + stats.wins() + "/" + stats.losses()
-        );
-        ko.computed(function () {
-          model.currentStar();
-          model.gwaioStats(
-            stats.turns() + "/" + stats.wins() + "/" + stats.losses()
-          );
+        var loadoutId = game.inventory().cards()[0].id;
+        model.gwaioLoadout = ko.observable("");
+        requireGW(["cards/" + loadoutId], function (card) {
+          model.gwaioLoadout(loc(card.summarize()));
         });
 
-        var originSystem = game
-          .galaxy()
-          .stars()
-          [game.galaxy().origin()].system();
+        var rgb = function (color) {
+          return "rgb(" + color[0] + "," + color[1] + "," + color[2] + ")";
+        };
+        var intelligence = function (subcommander) {
+          return {
+            name: subcommander.name,
+            color: rgb(
+              (subcommander.color && subcommander.color[0]) || [255, 255, 255]
+            ),
+            character: loc(subcommander.character),
+          };
+        };
+        model.gwaioPlayer = ko.computed(function () {
+          var inventory = game.inventory();
+          var playerName = ko.observable().extend({ session: "displayName" });
+          var playerColor = rgb(inventory.getTag("global", "playerColor")[0]);
 
-        if (originSystem.gwaio) {
-          model.gwaioVersion = originSystem.gwaio.version;
-          model.gwaioDifficulty = loc(originSystem.gwaio.difficulty);
-          model.gwaioSize = loc(originSystem.gwaio.galaxySize);
+          var commanders = [
+            { name: playerName, color: playerColor, character: loc("Human") },
+          ];
 
-          if (originSystem.gwaio.factionScaling === true)
-            model.gwaioFactionScaling = loc("!LOC:Enabled");
-          else model.gwaioFactionScaling = loc("!LOC:Disabled");
-
-          if (originSystem.gwaio.systemScaling === false) {
-            model.gwaioSystemScaling = loc("!LOC:Disabled");
-          } else {
-            model.gwaioSystemScaling = loc("!LOC:Enabled");
+          var subCommanders = inventory.minions();
+          if (subCommanders.length > 0) {
+            // eslint-disable-next-line lodash/prefer-map
+            _.forEach(subCommanders, function (subcommander) {
+              commanders.push(intelligence(subcommander));
+            });
           }
-
-          if (originSystem.gwaio.easierStart === true)
-            model.gwaioEasierStart = loc("!LOC:Enabled");
-          else model.gwaioEasierStart = loc("!LOC:Disabled");
-
-          if (originSystem.gwaio.tougherCommanders === true) {
-            model.gwaioTougherCommanders = loc("!LOC:Enabled");
-          } else {
-            model.gwaioTougherCommanders = loc("!LOC:Disabled");
-          }
-        } else {
-          model.gwaioVersion = loc("!LOC:Unknown");
-          model.gwaioDifficulty = loc("!LOC:Unknown");
-          model.gwaioSize = game.galaxy().stars().length;
-          model.gwaioFactionScaling = loc("!LOC:Unknown");
-          model.gwaioSystemScaling = loc("!LOC:Enabled");
-          model.gwaioEasierStart = loc("!LOC:Unknown");
-          model.gwaioTougherCommanders = loc("!LOC:Unknown");
-        }
-
-        if (originSystem.gwaio && originSystem.gwaio.ai) {
-          model.gwaioAI = originSystem.gwaio.ai;
-        } else model.gwaioAI = "Titans";
-
-        if (game.hardcore() === true) model.gwaioHardcore = loc("!LOC:Enabled");
-        else model.gwaioHardcore = loc("!LOC:Disabled");
+          return commanders;
+        });
 
         $("#header").append(
           loadHtml(